refactor: adjust package theme
