+++ app/src/main/java/com/app/desafiodourado/components/button/CustomButton.kt
@@ -18,9 +18,9 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.constraintlayout.compose.ConstraintLayout
 import com.app.desafiodourado.R
-import com.app.desafiodourado.ui.theme.BrowLight
-import com.app.desafiodourado.ui.theme.CustomDimensions
-import com.app.desafiodourado.ui.theme.Success
+import com.app.desafiodourado.theme.BrowLight
+import com.app.desafiodourado.theme.CustomDimensions
+import com.app.desafiodourado.theme.Success
 
 @Composable
 fun CustomButton(
