+++ app/src/main/java/com/app/desafiodourado/components/snackbar/SnackbarCustomType.kt
@@ -1,9 +1,9 @@
 package com.app.desafiodourado.components.snackbar
 
 import androidx.compose.ui.graphics.Color
-import com.app.desafiodourado.ui.theme.BrowDark
-import com.app.desafiodourado.ui.theme.Error
-import com.app.desafiodourado.ui.theme.Success
+import com.app.desafiodourado.theme.BrowDark
+import com.app.desafiodourado.theme.Error
+import com.app.desafiodourado.theme.Success
 
 enum class SnackbarCustomType(val background: Color, val text: Color) {
     SUCCESS(Success, Color.White),
